code refactoring, added db/users
