+++ router/router.go
@@ -2,7 +2,7 @@ package router
 
 import (
 	"github.com/TheDevExperiment/server/internal/auth"
-	"github.com/TheDevExperiment/server/internal/db/repositories"
+	"github.com/TheDevExperiment/server/internal/db/users"
 	"github.com/TheDevExperiment/server/middlewares"
 	"github.com/gin-gonic/gin"
 )
@@ -18,7 +18,7 @@ import (
 
 func SetupRouter() *gin.Engine {
 	r := gin.Default()
-	userRepository := repositories.NewUserRepository()
+	userRepository := users.NewUserRepository()
 
 	// TODO: register middleware
 	r.Use(func(c *gin.Context) {
